requirements.txt

python-dotenv
pynytimes
pandas
openpyxl



